+++ src/nvim/bufwrite.c
@@ -18,6 +18,7 @@
 #include "nvim/bufwrite.h"
 #include "nvim/change.h"
 #include "nvim/context.h"
+#include "nvim/dialog.h"
 #include "nvim/drawscreen.h"
 #include "nvim/errors.h"
 #include "nvim/eval/typval_defs.h"
@@ -30,7 +31,6 @@
 #include "nvim/globals.h"
 #include "nvim/highlight_defs.h"
 #include "nvim/iconv_defs.h"
-#include "nvim/input.h"
 #include "nvim/macros_defs.h"
 #include "nvim/mbyte.h"
 #include "nvim/memline.h"
