+++ src/nvim/dialog.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <stdbool.h>  // IWYU pragma: keep
+
+extern char *confirm_msg;      ///< ":confirm" message
+extern char *confirm_buttons;  ///< ":confirm" buttons sent to cmdline as prompt
+
+#include "dialog.h.generated.h"
