+++ src/nvim/ex_docmd.c
@@ -34,6 +34,7 @@
 #include "nvim/context.h"
 #include "nvim/cursor.h"
 #include "nvim/debugger.h"
+#include "nvim/dialog.h"
 #include "nvim/digraph.h"
 #include "nvim/drawscreen.h"
 #include "nvim/edit.h"
@@ -61,7 +62,6 @@
 #include "nvim/fold.h"
 #include "nvim/garray.h"
 #include "nvim/garray_defs.h"
-#include "nvim/getchar.h"
 #include "nvim/gettext_defs.h"
 #include "nvim/globals.h"
 #include "nvim/highlight_defs.h"
