+++ src/nvim/globals.h
@@ -8,8 +8,8 @@
 #include "nvim/event/loop.h"
 #include "nvim/ex_cmds_defs.h"
 #include "nvim/ex_eval_defs.h"
-#include "nvim/getchar_defs.h"
 #include "nvim/iconv_defs.h"
+#include "nvim/input_defs.h"
 #include "nvim/macros_defs.h"
 #include "nvim/menu_defs.h"
 #include "nvim/os/os_defs.h"
