+++ src/nvim/memory.c
@@ -836,8 +836,8 @@ char *arena_strdup(Arena *arena, const char *str)
 # include "nvim/ex_cmds.h"
 # include "nvim/ex_docmd.h"
 # include "nvim/file_search.h"
-# include "nvim/getchar.h"
 # include "nvim/grid.h"
+# include "nvim/input.h"
 # include "nvim/mark.h"
 # include "nvim/msgpack_rpc/channel.h"
 # include "nvim/option.h"
