+++ src/nvim/state.c
@@ -13,8 +13,8 @@
 #include "nvim/event/loop.h"
 #include "nvim/event/multiqueue.h"
 #include "nvim/ex_getln.h"
-#include "nvim/getchar.h"
 #include "nvim/globals.h"
+#include "nvim/input.h"
 #include "nvim/insexpand.h"
 #include "nvim/keycodes.h"
 #include "nvim/log.h"
