+++ src/nvim/tag.c
@@ -17,6 +17,7 @@
 #include "nvim/cmdexpand.h"
 #include "nvim/cmdexpand_defs.h"
 #include "nvim/cursor.h"
+#include "nvim/dialog.h"
 #include "nvim/drawscreen.h"
 #include "nvim/errors.h"
 #include "nvim/eval.h"
@@ -36,7 +37,6 @@
 #include "nvim/hashtab_defs.h"
 #include "nvim/help.h"
 #include "nvim/highlight_defs.h"
-#include "nvim/input.h"
 #include "nvim/insexpand.h"
 #include "nvim/lua/executor.h"
 #include "nvim/macros_defs.h"
